fix more gcc warnings

+++ src/base/init/init.c
@@ -281,7 +281,7 @@ void low_mem_init(void)
       /* switch on vm86-only JIT CPU emulation to with non-zero base */
       config.cpuemu = 3;
       init_emu_cpu();
-      c_printf("CONF: JIT CPUEMU set to 3 for %d86\n", vm86s.cpu_type);
+      c_printf("CONF: JIT CPUEMU set to 3 for %d86\n", (int)vm86s.cpu_type);
       error("Using CPU emulation because vm.mmap_min_addr > 0.\n"
 	      "You can most likely avoid this problem by running\n"
 	      "sysctl -w vm.mmap_min_addr=0\n"
